Pin MarkupSafe version to 2.0.1 (#48)
Back in February MarkupSafe released a new version of the package that does not include soft_unicode. Turns out Jinja has a dependency on MarkupSafe which is not pinned, so when MarkupSafe was updated it ended up breaking Jinja and subsequently dbt deps as shown by this community member.

This PR adds the hot-fix suggested in the thread, we should discuss with core what the long-term solution for this is.