Use contextmanager instead of pytest fixture
This is much more straightforward. We are not relying on dependency injection. DI is not inherently bad, but we lose type hinting when we go that route.
I think that using an explicit, immported context manager makes it easier to follow. Additionally, we will be able to pass arguments to test_context() in the future to customize the setup per test